requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.39.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
streamlit==1.28.1
requests==2.31.0
